% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sjr_journals}
\alias{sjr_journals}
\title{SCImago Journal Rank}
\format{
A tibble with 768,391 rows and 26 variables:
  \describe{
    \item{year}{Year of SCImago Journal Ranking calculation.}
    \item{rank}{Rank of the journal among all journals.}
    \item{sourceid}{Database ID of the journal.}
    \item{title}{Journal's title.}
    \item{type}{Type: "journal", "book series", "trade journal", or "conference and proceedings"}
    \item{issn}{ISSN journal identifier.}
    \item{publisher}{Publisher of the journal.}
    \item{open_acess}{A dummy variable encoding whether the journal is Open Acess.}
    \item{open_acess_diamond}{A dummy variable encoding whether the journal is **Diammond** Open Acess.}
    \item{sjr}{SCImago Journal Rank indicator. It expresses the average number of weighted citations received in the selected year by the documents published in the selected journal in the three previous years, --i.e. weighted citations received in year X to documents published in the journal in years X-1, X-2 and X-3. See [detailed description of SJR](https://www.scimagojr.com/SCImagoJournalRank.pdf) (PDF).}
    \item{sjr_best_quartile}{Highest quartile of the journal among all categories it belongs to.}
    \item{h_index}{Hirsch index of the journal. The h index expresses the journal's number of articles (h) that have received at least h citations. It quantifies both journal scientific productivity and scientific impact and it is also applicable to scientists, countries, etc. ([see H-index Wikipedia definition](http://en.wikipedia.org/wiki/Hirsch_number)).}
    \item{total_docs_year}{Total number of published documents within a specific year. All types of documents are considered, including citable and non citable documents.}
    \item{total_docs_3years}{Published documents in the three previous years (selected year documents are excluded), i.e.when the year X is selected, then X-1, X-2 and X-3 published documents are retrieved. All types of documents are considered, including citable and non citable documents.}
    \item{total_refs}{Total number of citations received by a journal to the documents published within a specific year.}
    \item{total_citations_3years}{Number of citations received in the selected year by a journal to the documents published in the three previous years, --i.e. citations received in year X to documents published in years X-1, X-2 and X-3. All types of documents are considered.}
    \item{citable_docs_3years}{Number of citable documents published by a journal in the three previous years (selected year documents are excluded). Exclusively articles, reviews and conference papers are considered.}
    \item{citations_doc_2years}{Average citations per document in a 2 year period. It is computed considering the number of citations received by a journal in the current year to the documents published in the two previous years, --i.e. citations received in year X to documents published in years X-1 and X-2. Comparable to Journal Impact Factor.}
    \item{ref_doc}{Average number of references per document in the selected year.}
    \item{percent_female}{Percentage of female authors. To determine the gender of authors, a method based on given names was employed, grounded in "cultural consensus theory" (See Van Buskirk et al. [2023](https://ojs.aaai.org/index.php/ICWSM/article/view/22195)).}
    \item{overton}{Number of documents cited by public policy documents according to [Overton](https://www.overton.io) database.}
    \item{sdg}{Number of documents related to the Sustainable Development Goals defined by the United Nations according to the [Elsevier 2023 Sustainable Development Goals (SDGs) Mapping](https://elsevier.digitalcommonsdata.com/datasets/y2zyy9vwzy/1). Available from 2018 onwards.}
    \item{country}{Country of the publisher.}
    \item{region}{Region of the publisher.}
    \item{coverage}{Years covered.}
    \item{categories}{Categories the journal belongs to.}
    \item{area}{Areas of science. Only appeared in 2022.}
  }
}
\source{
SCImago, (n.d.). SJR — SCImago Journal & Country Rank [Portal]. Retrieved 2026-06-05, from http://www.scimagojr.com
  \url{https://www.scimagojr.com/journalrank.php}
}
\usage{
sjr_journals
}
\description{
SCImago Journal Rank for all journals indexed by Scopus in 1999--2025
}
\examples{

\dontrun{
library(tidyverse)
library(sjrdata)

# Nature VS Science
sjr_journals \%>\%
    filter(title \%in\% c("Nature", "Science")) \%>\%
    ggplot(aes(cites_doc_2years, sjr, color = title))+
    geom_path(size = 1, alpha = .5)+
    geom_label(aes(label = year \%>\% str_sub(3, 4)),
               size = 3, label.padding = unit(.15, "line"))

# Demographic journals
sjr_journals \%>\%
    filter(title \%in\% c(
        "Demography",
        "Population and Development Review",
        "European Journal of Population",
        "Population Studies",
        "Demographic Research",
        "Genus"
    )) \%>\%
    ggplot(aes(cites_doc_2years, sjr, color = title))+
    geom_point()+
    stat_ellipse()+
    scale_color_brewer(palette = "Dark2")+
    coord_cartesian(expand = F)
}
}
\keyword{datasets}
